changed editprofile page, removed all details from signup page
Genres is now a dropdown select menu. Changed the name of string in Instrument collection. Removed instruments from musician collection (was for testing purposes). Updated MusiciansGenres to use dropdown menu for genres. Made schema for AddMusician page so it can deal with genres and instruments. Added new pages for schema. Editted typos in EditProfile. Removed all the information in signup.